backend/pyproject.toml
[tool.poetry]
name = "pfamserver"
version = "2.2.13"
description = "Mutual Information Server to Infer Coevolution"
authors = [
    "Eloy Colell <eloy.colell@gmail.com>",
    "Franco Simonetti <javiserte@gmail.com>",
    "Javier Iserte <javiserte@gmail.com>",
]
license = "MIT License"
classifiers = [
        'Development Status :: 5 - Production/Stable',
        'Topic :: Utilities',
        'Programming Language :: Python :: 3.8',
        'License :: OSI Approved :: GNU Affero General Public License v3'
]
homepage = "mistic2.leloir.org.ar:5001"

[tool.poetry.dependencies]
python = "^3.8"
cachetools = "^4.2.2"
click = "7.0"
Flask = "1.0.2"
environs = "^9.3.2"
enum34 = "1.1.4"
Flask-Caching = "1.10.1"
Flask-Collect = "^1.3.2"
Flask-DotEnv = "0.1.0"
Flask-Mail = "0.9.1"
flask-restplus = "^0.13.0"
Flask-SQLAlchemy = "2.4.4"
Flask-Webpack = "^0.1.0"
Flask-WTF = "^0.15.1"
merry = "^0.3.0"
python-dotenv = "^0.18.0"
requests = {extras = ["security"], version = "^2.25.1"}
SQLAlchemy = "1.3.24"
SQLAlchemy-Utils = "^0.37.8"
uWSGI = "^2.0.19"
Werkzeug = "0.16.0"
future = "^0.18.2"
greenlet = ">=0.4.17"
gevent = ">=20.9"
gunicorn = "^20.1.0"
types-six = "^0.1.7"
types-requests = "^2.25.1"
raven = "^6.10.0"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
bumpversion = "^0.6.0"
mypy = "^0.910"
pre-commit = "^2.13.0"
pycodestyle = "^2.7.0"
pydocstyle = "^6.1.1"
radon = "^5.0.1"
rope = "^0.19.0"
safety = "^1.10.3"
vim-debug = "^1.5.4"
virtualenv = "^20.4.7"
factory-boy = "^3.2.0"
Faker = "^8.10.1"
more-itertools = "^8.8.0"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pytest-html = "^3.1.1"
pytest-mock = "^3.6.1"
pytest-sugar = "^0.9.4"
pytest-xdist = "^2.3.0"
requests-mock = "^1.9.3"
responses = "^0.13.3"
tox = "^3.23.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
